Create hackerrank_data_structures.py
Another category of Problem Solving questions on hackerrank.com; with a total of 121 questions as of early 2021, it is worth having a separate file.

A very large (1000s of lines) .py file makes the kite linting application slow.  Kite is a code formatting review application which runs in the background and uses a lot of memory and compute resources for 'large' files, since the formatting is continually being processed.